requirements.txt
langgraph>=0.0.30
langchain-core>=0.1.30
langchain-openai>=0.1.0
langchain-anthropic>=0.1.0
pydantic>=2.0.0
python-dotenv>=1.0.0
